Move event monitor errors into their own module
